chore(menu) stop trying to cache the menu
The menu is cached but the elements are not. Every time
`createMenu` is called, which is often, new items are created.
This keeps adding new elements to the DOM instead of using the
existing ones. videojs always assumes createMenu will destroy the
old and create the new, so do that here too.